+++ packages/template-retail-react-app/app/components/field/index.jsx
@@ -39,7 +39,8 @@ const Field = ({
     autoComplete,
     defaultValue,
     helpText,
-    children
+    children,
+    inputRef
 }) => {
     const intl = useIntl()
     const [hidePassword, setHidePassword] = useState(true)
@@ -55,6 +56,7 @@ const Field = ({
           })
     const inputType =
         type === 'password' && hidePassword ? 'password' : type === 'password' ? 'text' : type
+
     return (
         <Controller
             name={name}
@@ -75,7 +77,10 @@ const Field = ({
                                 type
                             ) && (
                                 <Input
-                                    ref={ref}
+                                    ref={(node) => {
+                                        ref(node)
+                                        if (inputRef) inputRef.current = node
+                                    }}
                                     onChange={onChange}
                                     value={value}
                                     type={inputType}
@@ -175,7 +180,8 @@ Field.propTypes = {
     control: PropTypes.object,
     defaultValue: PropTypes.any,
     helpText: PropTypes.any,
-    children: PropTypes.any
+    children: PropTypes.any,
+    inputRef: PropTypes.object
 }
 
 export default Field